+++ src/rosros/patch.py
@@ -7,7 +7,7 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     23.02.2022
-@modified    05.12.2023
+@modified    07.12.2023
 ------------------------------------------------------------------------------
 """
 ## @namespace rosros.patch
@@ -59,6 +59,8 @@ def patch_ros1():
     if PATCHED: return
 
     rospy.Publisher.publish = Publisher__publish
+
+    rospy.Service.__init__ = Service__init
 
     rospy.ServiceProxy.call             = ServiceProxy__call
     rospy.ServiceProxy.call_async       = ServiceProxy__call_async
@@ -220,6 +222,14 @@ def set_extra_attribute(obj, name, value):
 
 if rospy:  # Patch-functions to apply on ROS1 classes, to achieve parity with ROS2 API
 
+    def service_serve_wrapper(self, serve):
+        """Returns service serve-function wrapped to ensure return with response instance."""
+        def inner(req):
+            resp = serve(req)
+            return {} if resp is None else resp
+        return functools.update_wrapper(inner, serve)
+
+
     def Publisher__assert_liveliness(self):
         """Does nothing (ROS2 compatibility stand-in)."""
 
@@ -231,6 +241,14 @@ if rospy:  # Patch-functions to apply on ROS1 classes, to achieve parity with RO
         attributes = functools.partial(ros.get_message_fields, self.data_class)
         msg = util.ensure_object(self.data_class, attributes, api.dict_to_message, *args, **kwds)
         return ROS1_Publisher__publish(self, msg)
+
+
+    ROS1_Service__init = rospy.Service.__init__
+
+    def Service__init(self, name, service_class, handler, buff_size=65536, error_handler=None):
+        """Wraps Service.__init__() to support returning None from server callback."""
+        callback = service_serve_wrapper(self, handler)
+        ROS1_Service__init(self, name, service_class, callback, buff_size, error_handler)
 
 
     ROS1_ServiceProxy__call = rospy.ServiceProxy.call
@@ -543,9 +561,11 @@ elif rclpy:  # Patch-functions to apply on ROS2 classes, to achieve parity with 
     def service_serve_wrapper(self, serve):
         """Returns service serve-function wrapped to ensure return with response instance."""
         def inner(req, resp):
-            resp, respcls, args, kwargs = serve(req), self.srv_type.Response, None, None
+            resp0, respcls, args, kwargs = resp, self.srv_type.Response, None, None
+            resp = serve(req, resp0)
             if   isinstance(resp, (list, tuple)): args   = resp
             elif isinstance(resp, dict):          kwargs = resp
+            elif resp is None:                    resp   = resp0
             if args is not None or kwargs is not None:
                 attributes = functools.partial(ros.get_message_fields, respcls)
                 args, kwargs = (args or []), (kwargs or {})
@@ -669,7 +689,7 @@ elif rclpy:  # Patch-functions to apply on ROS2 classes, to achieve parity with 
 
     def Service__init(self, service_handle, srv_type, srv_name, callback,
                       callback_group, qos_profile):
-        """Wraps Service.__init__() to support returning list or dict from server callback."""
+        """Wraps Service.__init__() to support returning list/dict/None from server callback."""
         callback = service_serve_wrapper(self, callback)
         ROS2_Service__init(self, service_handle, srv_type, srv_name, callback,
                            callback_group, qos_profile)