+++ CHANGELOG.rst
@@ -2,6 +2,19 @@
 Changelog for package rosros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.2.2 (2023-12-07)
+------------------
+* add rosros.wait_for_publisher(), wait_for_subscriber() and wait_for_service()
+* allow service callbacks to return None
+* fix service calls raising error in ROS2
+* fix rosros.get_services() raising error on node-parameter in ROS2
+* fix rosros.get_param_names() returning params dictionary instead of list
+* fix rosros.set_param() raising error on no descriptor in ROS2
+* fix util.get_arity() and wrap_arity() raising for built-in functions
+* fix rosros.spin_once() raising error on negative timeout in ROS1
+* fix api.get_message_type() not returning value for service in ROS1
+* fix throttling log messages again
+
 0.2.1 (2023-12-06)
 ------------------
 * add api.to_duration()